* Merged chroma key feature, but removed unsupported parameters and color names. Implemented chroma mode in image_transform instead of global layer setting, thus supporting animations as well. * Implemented support for chroma in scene_producer as well.
Former-commit-id: 54debe4164799c1a29134e4119a7b88cf56c0956 [formerly 54debe4164799c1a29134e4119a7b88cf56c0956 [formerly fa017fbf5fecf9479168e68d93abab552515047c]]
Former-commit-id: 8104a4058451ba23bdc38d9940f7abc30ef7afff
Former-commit-id: ffe5f0725e5deb3da9d3cb2f492f63867dcd78e1

+++ accelerator/ogl/util/shader.cpp
@@ -128,9 +128,9 @@ public:
 		GL(glUniform1f(get_location(name.c_str()), value));
 	}
 	
-	void set(const std::string& name, float value0, float value1)
+	void set(const std::string& name, double value0, double value1)
 	{
-		GL(glUniform2f(get_location(name.c_str()), value0, value1));
+		GL(glUniform2f(get_location(name.c_str()), static_cast<float>(value0), static_cast<float>(value1)));
 	}
 
 	void set(const std::string& name, double value)
@@ -149,7 +149,7 @@ shader::~shader(){}
 void shader::set(const std::string& name, bool value){impl_->set(name, value);}
 void shader::set(const std::string& name, int value){impl_->set(name, value);}
 void shader::set(const std::string& name, float value){impl_->set(name, value);}
-void shader::set(const std::string& name, float value0, float value1){impl_->set(name, value0, value1);}
+void shader::set(const std::string& name, double value0, double value1){impl_->set(name, value0, value1);}
 void shader::set(const std::string& name, double value){impl_->set(name, value);}
 int shader::id() const{return impl_->program_;}
 void shader::use()const{impl_->use();}
